deployment/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
prometheus-fastapi-instrumentator==7.0.0
kafka-python==2.0.2
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
httpx==0.27.0
tensorflow-cpu>=2.16.0
